Moved connection to pin 1 to pin A5. Added more slider controls. Increased encoder sensitivity.
Moved connection to pin 1 to pin A5 because Serial was probably interfereing with the charging of the inductor.

Added (barely) ability to differntiate between single and double clicks.
With the use of single clicks users can change the speed they move through a slide menu, and with a double click they exit it.
Fixed a couple of issues with button presses, such as one counting as many and the button remaining pressed in the program forever until the test function is called.

Increased encoder sensitivity to make it feel nicer to navigate through menus