Redid configuration system. Config now saves in permanent memory. Added documentation
Configuration now runs from data inside a struct.
LCDBrightness and LCDContrast are now bytes for efficency.
This struct is saved to EEPROM when written to. If no config has been written to EEPROM yet, it will initialize it with custom values. I recommend clearing the arduino's EEPROM before flashing on the program.